fix syntax

+++ app/api/ride-info/delete/[id]/route.tsx
@@ -5,6 +5,13 @@ import { NextRequest, NextResponse } from "next/server";
 export async function DELETE(request: NextRequest, { params }: { params: { id: string } }) {
     const { id } = params;
 
+    if (!id) {
+        return NextResponse.json(
+            { message: "ID parameter is missing" },
+            { status: 400 }
+        );
+    }
+
     await connectMongoDB();
     await RideInfo.findByIdAndDelete(id);
